move constant of derived values from optimize module to rewrite module
