+++ sharding-core/sharding-core-optimize/src/main/java/org/apache/shardingsphere/core/optimize/sharding/engnie/ShardingSelectOptimizeEngine.java
@@ -63,7 +63,7 @@ public final class ShardingSelectOptimizeEngine implements OptimizeEngine<Shardi
         for (SubqueryPredicateSegment each : subqueryPredicateSegments) {
             if (!each.getAndPredicates().isEmpty()) {
                 optimizedStatement.setContainsSubquery(true);
-                break;
+                return;
             }
         }
     }
